requirements.txt
# FastAPI server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# TTS - Chatterbox
chatterbox-tts>=0.1.2
torch>=2.1.0
torchaudio>=2.1.0

# Transcription - faster-whisper
faster-whisper>=1.1.0

# Audio processing
soundfile>=0.12.1
numpy>=1.26.4
librosa>=0.10.0

# Utilities
pydantic>=2.0.0
aiofiles>=24.1.0
